refactor deconvolve
