+++ biahub/cli/main.py
@@ -2,7 +2,6 @@ import click
 
 from biahub.cli.characterize_psf import characterize_psf
 from biahub.cli.concatenate import concatenate_cli
-from biahub.cli.deconvolve import deconvolve
 from biahub.cli.estimate_bleaching import estimate_bleaching
 from biahub.cli.estimate_deskew import estimate_deskew
 from biahub.cli.estimate_psf import estimate_psf
@@ -14,6 +13,7 @@ from biahub.cli.register import register
 from biahub.cli.segment import segment
 from biahub.cli.stabilize import stabilize
 from biahub.cli.stitch import stitch
+from biahub.deconvolve import deconvolve_cli
 from biahub.deskew import deskew_cli
 
 CONTEXT = {"help_option_names": ["-h", "--help"]}
@@ -42,6 +42,6 @@ cli.add_command(concatenate_cli)
 cli.add_command(estimate_stabilization)
 cli.add_command(stabilize)
 cli.add_command(estimate_psf)
-cli.add_command(deconvolve)
+cli.add_command(deconvolve_cli)
 cli.add_command(characterize_psf)
 cli.add_command(segment)
